+++ bin/09_init.sh
@@ -1,1 +1,0 @@
-sudo -u steam sh -c "${STEAM_HOME}/bin/killinuxfloor reset"