+++ lib/aliases.sh
@@ -1,1 +1,0 @@
-export ECHO_DONE='echo -e \e[32mdone\e[0m.'