+++ lib/globals.sh
@@ -1,3 +1,0 @@
-export STEAM_HOME='/home/steam'
-export SKIP_KFGAME=0
-export KF2_DEBUG=0