+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## killinuxfloor 2.0 (TBD)
+
+* Port installer and uninstaller to Ansible
+* Add support for CentOS 8
+* Add `killinuxfloor apply` command
+* Update Node.js to 12
+
 ## killinuxfloor 1.1 (2019-05-15)
 
 * Add support for bans via `killinuxfloor ban` command